+++ src/App.css
@@ -100,7 +100,7 @@ p {
 }
 
 .country-info:nth-child(2n + 1) {
-  background-color: var(--color-primary-dark);
+  background-color: var(--color-primary-light);
 }
 
 @media (max-width: 768px) {
