+++ src/components/ContriesList.jsx
@@ -1,0 +1,13 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+import Country from './Country';
+
+const ContriesList = ({ countries }) => (
+  <div className="countries-grid">
+    {countries.map((countryData) => (
+      <Country data={countryData} key={countryData.name} />
+    ))}
+  </div>
+);
+
+export default ContriesList;
