+++ src/components/DetailsPage.jsx
